Restructured Settings components to match the rest of the project
