+++ src/components/Settings/CreateLabel/CreateLabel.module.scss
@@ -1,0 +1,13 @@
+@import '../../../styles/variables.scss';
+
+.labelDesign {
+  display: inline-block;
+  color: white;
+  text-align: center;
+  padding-top: 2px;
+  margin-bottom: 10px;
+  width: 100px;
+  height: 20px;
+  border-radius: 25px;
+  font-size: 0.8rem;
+}
