+++ src/components/Settings/LabelList/LabelList.module.scss
@@ -1,0 +1,6 @@
+@import '../../../styles/variables.scss';
+
+.labelListStyle {
+  position: relative;
+  height: 25px;
+}
